requirements.txt
aiohttp==0.15.1
appdirs==1.4.0
asyncio==3.4.3
ConfigArgParse==0.9.3
Django==1.8.3
django-filter==0.10.0
djangorestframework==3.1.3
enum34==1.0.4
hangups==0.2.10
hangups-urwid==1.2.2.dev0
Markdown==2.6.2
purplex==0.2.4
readlike==0.1.1
ReParser==1.4.2
requests==2.6.0
six==1.9.0
